+++ jobmon_gui/src/configs/ExternalUrls.ts
@@ -1,3 +1,3 @@
 export const imageLauncherUrl = "https://imagelauncher.scicomp.ihme.washington.edu/"
 export const metricMindUrl = "https://ai.ihme.washington.edu/metricmind/chat"
-export const squidUrl = "https://squid.ihme.washington.edu"
+export const jobInfoUrl = "https://slurmtool.ihme.washington.edu/jobinfo"